WIP remove redux from keypad
